Makefile: build precomp generators using build-system toolchain
When cross-compiling libsecp256k1, if the `precomputed_ecmult*.c` source
files need to be regenerated, then the generators need to be built for
the *build* system, not for the *host* system. Autoconf supports this
fairly cleanly via the `AX_PROG_CC_FOR_BUILD` macro (from Autoconf
Archive), but Automake requires some hackery. When building the
generators, we override the `CC` variable to its build-system
counterpart, `CC_FOR_BUILD`, and we specify Automake per-program
overrides for `CFLAGS`, `CPPFLAGS`, and `LDFLAGS`, setting their values
respectively from the `CFLAGS_FOR_BUILD`, `CPPFLAGS_FOR_BUILD`, and
`LDFLAGS_FOR_BUILD` variables, whose values in turn are populated by the
aforementioned Autoconf macro and may be overridden on the `make`
command line. Since Automake lacks support for overriding `EXEEXT` on a
per-program basis, we define a recipe that builds the generator binaries
with names suffixed with `$(EXEEXT)` and then renames them suffixed with
`$(BUILD_EXEEXT)`.